+++ pkg/handlers/handlers.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"github.com/cpwu/nova/pkg/config"
+	"github.com/cpwu/nova/pkg/models"
 	"github.com/cpwu/nova/pkg/render"
 )
 
@@ -29,10 +30,17 @@ func NewHandlers(r *Repository) {
 
 // Home page handler that writes a simple response to the client and handles any potential errors that may occur during the writing process. It also prints the number of bytes written to the console for debugging purposes.
 func (m *Repository) Home(w http.ResponseWriter, r *http.Request) {
-	render.RenderTemplate(w, "home.page.tmpl")
+	render.RenderTemplate(w, "home.page.tmpl", &models.TemplateData{})
 }
 
 // About page handler that demonstrates the use of a helper function to add two values and print the result. It also writes a response to the client and handles any potential errors that may occur during the writing process.
 func (m *Repository) About(w http.ResponseWriter, r *http.Request) {
-	render.RenderTemplate(w, "about.page.tmpl")
+	// perform some logic
+	stringMap := make(map[string]string)
+	stringMap["test"] = "Hello again"
+
+	// send the data to the template
+	render.RenderTemplate(w, "about.page.tmpl", &models.TemplateData{
+		StringMap: stringMap,
+	})
 }